fix: reorganized card components
